refactor(types): 抽取 types 层，移动 models 和 exceptions
- core/models/ → types/models/ (8 个数据类)
- core/exceptions/ → types/exceptions/ (异常类型 + 错误码)
- 全局更新 import 路径
