+++ src/danmaku_sender/core/engines/editor_session.py
@@ -4,8 +4,8 @@ import colorsys
 from typing import Any, Callable
 
 from danmaku_sender.core.config import ValidationConfig
-from danmaku_sender.core.models.danmaku import Danmaku
-from danmaku_sender.core.models.editor_types import EditorItem, EditorField, AtomicChange, ViewItem, InsertPosition
+from danmaku_sender.types.models.danmaku import Danmaku
+from danmaku_sender.types.models.editor_types import EditorItem, EditorField, AtomicChange, ViewItem, InsertPosition
 from danmaku_sender.core.services.danmaku_validator import validate_danmaku_list
