+++ src/danmaku_sender/core/engines/sender/executor.py
@@ -2,11 +2,11 @@ import logging
 from threading import Event
 
 from danmaku_sender.api.bili_api_client import BiliApiClient
-from danmaku_sender.core.models.danmaku import Danmaku
-from danmaku_sender.core.models.common import VideoTarget
-from danmaku_sender.core.models.result import DanmakuSendResult
-from danmaku_sender.core.exceptions.exceptions import BiliApiError, BiliNetworkError
-from danmaku_sender.core.exceptions.api_errors import BiliDmErrorCode
+from danmaku_sender.types.models.danmaku import Danmaku
+from danmaku_sender.types.models.common import VideoTarget
+from danmaku_sender.types.models.result import DanmakuSendResult
+from danmaku_sender.types.exceptions.exceptions import BiliApiError, BiliNetworkError
+from danmaku_sender.types.exceptions.api_errors import BiliDmErrorCode
 
 
 class DanmakuExecutor:
