+++ src/danmaku_sender/core/engines/sender/scheduler.py
@@ -6,7 +6,7 @@ from .context import SendingContext, DanmakuFingerprint, SendJob
 from .delay_manager import DelayManager
 
 from danmaku_sender.core.database.history_manager import HistoryManager
-from danmaku_sender.core.models.danmaku import Danmaku
+from danmaku_sender.types.models.danmaku import Danmaku
 
 
 class DanmakuScheduler:
