+++ src/danmaku_sender/core/services/danmaku_validator.py
@@ -1,6 +1,6 @@
 from typing import TypedDict
 
-from danmaku_sender.core.models.danmaku import Danmaku
+from danmaku_sender.types.models.danmaku import Danmaku
 from danmaku_sender.core.config import ValidationConfig
