+++ src/danmaku_sender/runtime/account_manager.py
@@ -7,7 +7,7 @@ from cryptography.fernet import Fernet, InvalidToken
 from platformdirs import user_data_dir
 
 from ..config.app_config import AppInfo
-from ..core.models.account import AccountCredential
+from ..types.models.account import AccountCredential
 from .app_state import AppState
