+++ src/danmaku_sender/types/exceptions/api_errors.py
