+++ src/danmaku_sender/types/exceptions/exceptions.py
