+++ src/danmaku_sender/ui/controllers/monitor_controller.py
@@ -5,7 +5,7 @@ from PySide6.QtCore import QObject, Signal, Slot
 
 from danmaku_sender.api.bili_api_client import BiliApiClient
 from danmaku_sender.core.engines.bili_monitor import BiliDanmakuMonitor
-from danmaku_sender.core.models.common import VideoTarget
+from danmaku_sender.types.models.common import VideoTarget
 from danmaku_sender.core.config import ApiAuthConfig, MonitorConfig
 from danmaku_sender.ui.framework.concurrency import WorkerThread
 from danmaku_sender.utils.system_utils import KeepSystemAwake
