+++ src/danmaku_sender/ui/controllers/video_controller.py
@@ -3,7 +3,7 @@ import logging
 from PySide6.QtCore import QObject, Signal, QThreadPool, Slot
 
 from danmaku_sender.api.bili_api_client import BiliApiClient
-from danmaku_sender.core.models.video import VideoInfo
+from danmaku_sender.types.models.video import VideoInfo
 from danmaku_sender.core.config import ApiAuthConfig
 from danmaku_sender.core.services.video_fetcher import VideoFetcher
 from danmaku_sender.ui.framework.concurrency import PoolTask
