+++ src/danmaku_sender/ui/history/page.py
@@ -10,8 +10,8 @@ from PySide6.QtGui import QAction, QDesktopServices
 from .components import Col, HistoryTableModel
 from .dialogs import DanmakuDetailDialog
 
-from danmaku_sender.core.models.common import DanmakuStatus
-from danmaku_sender.core.models.video import VideoInfo
+from danmaku_sender.types.models.common import DanmakuStatus
+from danmaku_sender.types.models.video import VideoInfo
 from danmaku_sender.runtime.app_state import AppState
 from danmaku_sender.ui.controllers.video_controller import VideoController
 from danmaku_sender.ui.controllers.history_controller import HistoryController
